[WEB-1236] DRY out clinic profile form to share between clinic details and profile components
